Rebase from reproduce (#116)
* add windowed gs, with some hard coded part

* add window

* add window script

* bugfix

* add scripts for window experiment

* add uncommitted files

* update scripts

* update

* update GSW generator to avoid duplicate write under the same transaction

* update

* add stock

* update

* update dataset and figure scripts

* update stock exchange analysis experiment

* code clean and bugfix

* Add Reproduce scripts

* Add Reproduce scripts

* Add Reproduce scripts

* Update dataRootPath
Add project_Dir and jar_Dir

* Update dataRootPath
Add project_Dir and jar_Dir

* Update dataRootPath

* add flink performance comparison

* Update README.md

* rebase to SIGMOD23_Reproduce

* code clean

---------

+++ application/src/main/java/common/bolts/transactional/sl/SLBolt.java
@@ -18,8 +18,7 @@ import static common.CONTROL.enable_app_combo;
 import static common.CONTROL.enable_latency_measurement;
 import static common.Constants.DEFAULT_STREAM_ID;
 import static content.common.CommonMetaTypes.AccessType.READ_WRITE;
-import static profiler.MeasureTools.BEGIN_POST_TIME_MEASURE;
-import static profiler.MeasureTools.END_POST_TIME_MEASURE;
+import static profiler.MeasureTools.*;
 
 public abstract class SLBolt extends TransactionalBolt {
     SINKCombo sink;
@@ -31,7 +30,7 @@ public abstract class SLBolt extends TransactionalBolt {
     }
 
     @Override
-    protected void TXN_PROCESS(double _bid) throws DatabaseException, InterruptedException {
+    protected void TXN_PROCESS(long _bid) throws DatabaseException, InterruptedException {
     }
 
     protected void DEPOSITE_REQUEST_NOLOCK(DepositEvent event, TxnContext txnContext) throws DatabaseException {
@@ -126,9 +125,9 @@ public abstract class SLBolt extends TransactionalBolt {
     }
 
     //post stream processing phase..
-    protected void POST_PROCESS(double _bid, long timestamp, int combo_bid_size) throws InterruptedException {
+    protected void POST_PROCESS(long _bid, long timestamp, int combo_bid_size) throws InterruptedException {
         BEGIN_POST_TIME_MEASURE(thread_Id);
-        for (double i = _bid; i < _bid + combo_bid_size; i++) {
+        for (long i = _bid; i < _bid + combo_bid_size; i++) {
             if (input_event instanceof DepositEvent) {
                 ((DepositEvent) input_event).setTimestamp(timestamp);
                 DEPOSITE_REQUEST_POST((DepositEvent) input_event);
@@ -160,6 +159,6 @@ public abstract class SLBolt extends TransactionalBolt {
         }
     }
 
-    protected void LAL_PROCESS(double _bid) throws InterruptedException, DatabaseException {
+    protected void LAL_PROCESS(long _bid) throws InterruptedException, DatabaseException {
     }
 }
