+++ application/src/main/java/common/bolts/transactional/tp/TPBolt_LA.java
@@ -13,19 +13,19 @@ public abstract class TPBolt_LA extends TPBolt {
         super(log, fid, sink);
     }
 
-    protected void LAL(LREvent event, double i, double _bid) throws DatabaseException {
+    protected void LAL(LREvent event, long i, long _bid) throws DatabaseException {
         REQUEST_LOCK_AHEAD(event, txn_context[(int) (i - _bid)]);
     }
 
     //lock_ratio-ahead phase.
     @Override
-    protected void LAL_PROCESS(double _bid) throws DatabaseException, InterruptedException {
+    protected void LAL_PROCESS(long _bid) throws DatabaseException, InterruptedException {
         int _combo_bid_size = 1;
         BEGIN_WAIT_TIME_MEASURE(thread_Id);
         //ensures that locks are added in the input_event sequence order.
         transactionManager.getOrderLock().blocking_wait(_bid);
         long lock_time_measure = 0;
-        for (double i = _bid; i < _bid + _combo_bid_size; i++) {
+        for (long i = _bid; i < _bid + _combo_bid_size; i++) {
             txn_context[(int) (i - _bid)] = new TxnContext(thread_Id, this.fid, i);
             LREvent event = (LREvent) input_event;
             LAL(event, i, _bid);
@@ -36,10 +36,10 @@ public abstract class TPBolt_LA extends TPBolt {
         END_WAIT_TIME_MEASURE(thread_Id);
     }
 
-    protected void PostLAL_process(double _bid) throws DatabaseException {
+    protected void PostLAL_process(long _bid) throws DatabaseException {
         int _combo_bid_size = 1;
         //txn process phase.
-        for (double i = _bid; i < _bid + _combo_bid_size; i++) {
+        for (long i = _bid; i < _bid + _combo_bid_size; i++) {
             LREvent event = (LREvent) input_event;
             TXN_REQUEST_NOLOCK(event, txn_context[(int) (i - _bid)]);
             BEGIN_ACCESS_TIME_MEASURE(thread_Id);
